move apple simd stuff to separate header
